Split stats into their own modules
Weʼve seen memory usage slowly balloon in the stats component when new
modules pull in new dependencies.  The best pattern for keeping mem
usage low seems to be to keep the stats definitions in separate modules,
which this PR implements for mite core.
