+++ setup.cfg
@@ -24,8 +24,8 @@ console_scripts =
     mite = mite.__main__:main
 mite_stats =
     mite = mite.stats:_MITE_STATS
-    mite_http = mite_http:_MITE_STATS
-    mite_selenium = mite_selenium:_MITE_STATS
+    mite_http = mite_http.stats:STATS
+    mite_selenium = mite_selenium.stats:STATS
 
 [options.packages.find]
 exclude =
